Change highscore to double. Reorder MenuStates.

+++ Assets/BibaFramework/BibaGame/Scripts/Model/BibaGameModel.cs
@@ -18,7 +18,7 @@ namespace BibaFramework.BibaGame
         public bool HelpBubblesEnabled { get; set; }
         public bool ChartBoostEnabled { get; set; }
 
-        public float HighScore { get; set; }
+        public double HighScore { get; set; }
 
         public BibaGameModel()
         {
